Bug 1967790: Check all extra bits in alignment flags. r=rhunt
Because the alignment in a memarg is a LEB128 number, we actually should
check all the extra "flag" bits to make sure that they are zero. Because
we were immediately converting to a uint8_t we were losing potentially
three bytes of flags. This caused one spec test to fail.

Differential Revision: https://phabricator.services.mozilla.com/D251036

+++ js/src/wasm/WasmOpIter.h
@@ -1977,7 +1977,7 @@ inline bool OpIter<Policy>::readLinearMemoryAddress(
 
   uint8_t alignLog2 = flags & ((1 << 6) - 1);
   uint8_t hasMemoryIndex = flags & (1 << 6);
-  uint8_t undefinedBits = flags & ~((1 << 7) - 1);
+  uint32_t undefinedBits = flags & ~((1 << 7) - 1);
 
   if (undefinedBits != 0) {
     return fail("invalid memory flags");
